Issue #164 - Renamed most trains
- Slight rebalance for petrotrain
- Reorder of smeltingtrain
