+++ angelsaddons-crawlertrain/data.lua
@@ -1,8 +1,8 @@
 angelsmods = angelsmods or {}
 angelsmods.addons = angelsmods.addons or {}
 angelsmods.addons.crawlertrain = angelsmods.addons.crawlertrain or {}
-angelsmods.addons.crawlertrain.enable_tiers = settings.startup["angels-crawlertrain-tier-enable"].value
 angelsmods.addons.crawlertrain.tier_amount = settings.startup["angels-crawlertrain-tier-amount"].value
+angelsmods.addons.crawlertrain.enable_tiers = angelsmods.addons.crawlertrain.tier_amount > 1 and true or false
 angelsmods.addons.crawlertrain.number_tint = {r = 0.95, g = 0.65, b = 0.25, a = 1}
 
 tech_unlocks = {}
