+++ angelsaddons-crawlertrain/info.json
@@ -1,6 +1,6 @@
 {
   "name": "angelsaddons-crawlertrain",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "factorio_version": "0.18",
   "title": "Angel's Addons - Crawler Train",
   "author": "Arch666Angel",
