+++ angelsaddons-crawlertrain/locale/en/crawlertrain.cfg
@@ -1,20 +1,20 @@
 [entity-name]
-crawler-locomotive="Crawler" Construction Train Locomotive
-crawler-locomotive-wagon="Crawler" Construction Train Locomotive Wagon
-crawler-bot-wagon="Crawler" Construction Train Robot Wagon
-crawler-wagon="Crawler" Construction Train Wagon
+crawler-locomotive=Crawler Locomotive
+crawler-locomotive-wagon=Crawler Locomotive Tender
+crawler-bot-wagon=Crawler Robot Wagon
+crawler-wagon=Crawler Cargo Wagon
 
 [item-name]
-crawler-locomotive="Crawler" Construction Train Locomotive
-crawler-locomotive-wagon="Crawler" Construction Train Locomotive Wagon
-crawler-bot-wagon="Crawler" Construction Train Robot Wagon
-crawler-wagon="Crawler" Construction Train Wagon
+crawler-locomotive=Crawler Locomotive
+crawler-locomotive-wagon=Crawler Locomotive Tender
+crawler-bot-wagon=Crawler Robot Wagon
+crawler-wagon=Crawler Cargo Wagon
 
 [recipe-name]
-crawler-locomotive="Crawler" Construction Train Locomotive
-crawler-locomotive-wagon="Crawler" Construction Train Locomotive Wagon
-crawler-bot-wagon="Crawler" Construction Train Robot Wagon
-crawler-wagon="Crawler" Construction Train Wagon
+crawler-locomotive=Crawler Locomotive
+crawler-locomotive-wagon=Crawler Locomotive Tender
+crawler-bot-wagon=Crawler Robot Wagon
+crawler-wagon=Crawler Cargo Wagon
 
 [recipe-category-name]
 
@@ -23,8 +23,10 @@ crawler-wagon="Crawler" Construction Train Wagon
 [item-subgroup-name]
 
 [technology-name]
-angels-crawler-train="Crawler" Construction Train
+angels-crawler-train=Crawler Train
+
+[technology-description]
+angels-crawler-train=Crawler Construction trains
 
 [mod-setting-name]
-angels-crawlertrain-tier-enable=Enable higher tier trains
 angels-crawlertrain-tier-amount=Amount of tiers
