+++ angelsaddons-petrotrain/info.json
@@ -1,6 +1,6 @@
 {
   "name": "angelsaddons-petrotrain",
-  "version": "0.5.3",
+  "version": "0.5.4",
   "factorio_version": "0.18",
   "title": "Angel's Addons - Petrochem Train",
   "author": "Arch666Angel",
