+++ angelsaddons-petrotrain/locale/en/petrotrain.cfg
@@ -1,17 +1,17 @@
 [entity-name]
 petro-locomotive-1=PetroChem Locomotive
-petro-tank1=Petrochem Gas Tank Wagon
-petro-tank2=Petrochem Oil Tank Wagon
+petro-tank1=PetroChem Gas Fluid Wagon
+petro-tank2=PetroChem Oil Fluid Wagon
 
 [item-name]
 petro-locomotive-1=PetroChem Locomotive
-petro-tank1=Petrochem Gas Tank Wagon
-petro-tank2=Petrochem Oil Tank Wagon
+petro-tank1=PetroChem Gas Fluid Wagon
+petro-tank2=PetroChem Oil Fluid Wagon
 
 [recipe-name]
 petro-locomotive-1=PetroChem Locomotive
-petro-tank1=Petrochem Gas Tank Wagon
-petro-tank2=Petrochem Oil Tank Wagon
+petro-tank1=PetroChem Gas Fluid Wagon
+petro-tank2=PetroChem Oil Fluid Wagon
 
 [recipe-category-name]
 
@@ -23,5 +23,4 @@ petro-tank2=Petrochem Oil Tank Wagon
 angels-petro-train=PetroChem Trains
 
 [mod-setting-name]
-angels-petrotrain-tier-enable=Enable higher tier trains
 angels-petrotrain-tier-amount=Amount of tiers
