+++ angelsaddons-petrotrain/prototypes/entities/petrotrain.lua
@@ -284,7 +284,7 @@ funcs.generate_train_entities(
     minable = {mining_time = 1, result = "petro-tank1"},
     mined_sound = {filename = "__core__/sound/deconstruct-medium.ogg"},
     max_health = 600,
-    capacity = 25000,
+    capacity = mods.angelspetrochem and 40000 or 25000,
     corpse = "medium-remnants",
     dying_explosion = "medium-explosion",
     collision_box = {{-0.6, -2.4}, {0.6, 2.4}},
@@ -453,7 +453,7 @@ funcs.generate_train_entities(
     minable = {mining_time = 1, result = "petro-tank2"},
     mined_sound = {filename = "__core__/sound/deconstruct-medium.ogg"},
     max_health = 600,
-    capacity = 25000,
+    capacity = mods.angelspetrochem and 40000 or 25000,
     corpse = "medium-remnants",
     dying_explosion = "medium-explosion",
     collision_box = {{-0.6, -2.4}, {0.6, 2.4}},
