+++ angelsaddons-smeltingtrain/info.json
@@ -1,6 +1,6 @@
 {
   "name": "angelsaddons-smeltingtrain",
-  "version": "0.3.3",
+  "version": "0.3.4",
   "factorio_version": "0.18",
   "title": "Angel's Addons - Smelting Train",
   "author": "Arch666Angel",
