+++ angelsaddons-smeltingtrain/locale/en/smeltingtrain.cfg
@@ -1,17 +1,17 @@
 [entity-name]
 smelting-locomotive-1=Smelting Locomotive
 smelting-locomotive-tender=Smelting Locomotive Tender
-smelting-wagon-1=Smelting Wagon
+smelting-wagon-1=Smelting Cargo Wagon
 
 [item-name]
 smelting-locomotive-1=Smelting Locomotive
 smelting-locomotive-tender=Smelting Locomotive Tender
-smelting-wagon-1=Smelting Wagon
+smelting-wagon-1=Smelting Cargo Wagon
 
 [recipe-name]
 smelting-locomotive-1=Smelting Locomotive
 smelting-locomotive-tender=Smelting Locomotive Tender
-smelting-wagon-1=Smelting Wagon
+smelting-wagon-1=Smelting Cargo Wagon
 
 [recipe-category-name]
 
@@ -23,5 +23,4 @@ smelting-wagon-1=Smelting Wagon
 angels-smelting-train=Smelting Trains
 
 [mod-setting-name]
-angels-smeltingtrain-tier-enable=Enable higher tier trains
 angels-smeltingtrain-tier-amount=Amount of tiers
